Fix build with pointer-typed NULL (#3147)
If we switch NULL to be pointer-typed (rather than a plain int zero, as
it is currently), this reveals some improper usage of NULL where it is
being assigned to a non-pointer lvalue. We can't switch the type of NULL
as this breaks some matches, but we can fix these improper uses.

(cherry picked from commit b1bb205f78ebf3c7eafac7531e48d91b02184fb5)

+++ src/melee/ef/efsync.c
@@ -86,7 +86,7 @@ void* efSync_Spawn(s32 gfx_id, HSD_GObj* gobj, ...)
     efLib_LoadKind = EF_LOADKIND_ASYNC;
     efLib_AnimCount = 0;
     va_start(vlist, gobj);
-    if ((gfx_id == 0x479) && ((u32) efAsync_DatEntries[1].data == NULL)) {
+    if ((gfx_id == 0x479) && (efAsync_DatEntries[1].data == NULL)) {
         gfx_id = 0x506;
     }
     if (gfx_id < 0x250) {
